S35途中経過
LEDとstdio_init_all入れた

+++ testcode/s35/s35.cpp
@@ -37,7 +37,7 @@ void S35::stop() {
 
 void S35::left_turn() {
     // パルス幅を設定(1000[us]~2000[us]で動作するらしい)
-    double pulse_width = 2000; //[us]
+    int pulse_width = 2000; //[us]
 
     uint slice_num = pwm_gpio_to_slice_num(pin);
 
@@ -52,7 +52,7 @@ void S35::left_turn() {
 void S35::right_turn() {
 
     // パルス幅を設定(1000[us]~2000[us]で動作するらしい)
-    double pulse_width = 1000;//[us]
+    int pulse_width = 1000;//[us]
 
     uint slice_num = pwm_gpio_to_slice_num(pin);
 
@@ -67,7 +67,7 @@ void S35::right_turn() {
 void S35::stop_turn() {
 
     // パルス幅を設定(1000[us]~2000[us]で動作するらしい)
-    double pulse_width = 1500;//[us]
+    int pulse_width = 1500;//[us]
 
     uint slice_num = pwm_gpio_to_slice_num(pin);
